requirements.txt
ariadne
flask
sqlalchemy
passlib[bcrypt]
python-jose
python-dotenv    